# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.CreateUserRequest:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      user_name:
        type: string
    required:
    - date_of_birth
    - email
    - password
    - user_name
    type: object
...
  github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.EditUserRequest:
    properties:
      cover_picture:
        type: string
      date_of_birth:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      profile_picture:
        type: string
    type: object
  github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
...
  github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.LoginRequest:
    properties:
      password:
        type: string
      user_name:
        type: string
    required:
    - password
    - user_name
    type: object
  github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.LoginResponse:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.UserDetailInfo'
    type: object
  github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.MessageResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.UserDetailInfo:
    properties:
      cover_picture:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      profile_picture:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
    type: object
  github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.UserDetailInfoResponse:
    properties:
      cover_picture:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      profile_picture:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
    type: object
...
host: localhost:19003
info:
  contact:
    email: support@wandersphere.com
    name: API Support
  description: API Server for WanderSphere Application
  title: WanderSphere API
  version: "1.0"
paths:
...
  /users/{user_id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.UserDetailInfoResponse'
        "400":
          description: Invalid user ID or user not found
          schema:
            $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.MessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.MessageResponse'
      summary: Get user details
      tags:
      - users
  /users/edit:
    post:
      consumes:
      - application/json
      description: Update user profile information
      parameters:
      - description: User information to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.EditUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.MessageResponse'
        "400":
          description: Validation error or user not found
          schema:
            $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.MessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Edit user profile
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Login with username and password
      parameters:
      - description: Login credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.LoginResponse'
        "400":
          description: Validation error or authentication failed
          schema:
            $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.ErrorResponse'
      summary: Authenticate a user
      tags:
      - users
  /users/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User created successfully
          schema:
            $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.MessageResponse'
        "400":
          description: Validation error or user already exists
          schema:
            $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.MessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_hoangNguyenDev3_WanderSphere_backend_internal_pkg_types.MessageResponse'
      summary: Register a new user
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"